pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "multi-armed-bandit-framework"
version = "0.1.0"
description = "Production-ready, containerized Multi-Armed Bandit experimentation platform (FastAPI + Redis + Streamlit)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "numpy>=1.23",
]

[project.optional-dependencies]
api = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "redis[hiredis]>=5.0",
  "pydantic-settings>=2.2",
  "mangum>=0.17",
]
bench = [
  "matplotlib>=3.7",
]
dashboard = [
  "streamlit>=1.30",
  "requests>=2.31",
  "plotly>=5.18",
]
dev = [
  "pytest>=7.4",
  "httpx>=0.25",
  "ruff>=0.2",
  "mypy>=1.8",
  "types-requests>=2.31",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
strict_equality = true
pretty = true

[tool.pytest.ini_options]
testpaths = ["tests"]
